# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/tasks": {
      "post": {
        "description": "Creates a new task. Only admins can access this endpoint.",
        "operationId": "TaskController_createTask",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "User not authenticated."
          },
          "403": {
            "description": "Only admin can create tasks."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Create a new task (Admin)",
        "tags": [
          "Task"
        ]
      },
      "get": {
        "description": "Returns a paginated list of all tasks. Admin only.",
        "operationId": "TaskController_getAllTasks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Tasks fetched successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTasksResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "User not authenticated."
          },
          "403": {
            "description": "Only admin can view all tasks."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Get all tasks (Admin)",
        "tags": [
          "Task"
        ]
      }
    },
    "/tasks/{taskId}/assign/{userId}": {
      "post": {
        "description": "Assigns a task to a specific user. Admin only.",
        "operationId": "TaskController_assignTask",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Task assigned successfully."
          },
          "400": {
            "description": "Invalid task or user ID."
          },
          "401": {
            "description": "User not authenticated."
          },
          "403": {
            "description": "Only admin can assign tasks."
          },
          "404": {
            "description": "Task or user not found."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Assign a task to a user (Admin)",
        "tags": [
          "Task"
        ]
      }
    },
    "/tasks/{taskId}/status": {
      "patch": {
        "description": "Updates the status of a task assigned to the logged-in user.",
        "operationId": "TaskController_updateTaskStatus",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssignedTaskStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task status updated successfully."
          },
          "400": {
            "description": "Invalid status value."
          },
          "401": {
            "description": "User not authenticated."
          },
          "403": {
            "description": "Task not assigned to this user."
          },
          "404": {
            "description": "Task not found."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Update task status",
        "tags": [
          "Task"
        ]
      },
      "get": {
        "description": "Returns the current status of a task. Admin only.",
        "operationId": "TaskController_GetTaskStatus",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task status fetched successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateAssignedTaskStatusDto"
                }
              }
            }
          },
          "401": {
            "description": "User not authenticated."
          },
          "403": {
            "description": "Only admin can view this status."
          },
          "404": {
            "description": "Task not found."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Get task status (Admin)",
        "tags": [
          "Task"
        ]
      }
    },
    "/tasks/{id}": {
      "put": {
        "description": "Updates task details by ID. Admin only.",
        "operationId": "TaskController_updateTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "User not authenticated."
          },
          "403": {
            "description": "Only admin can update tasks."
          },
          "404": {
            "description": "Task not found."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Update a task (Admin)",
        "tags": [
          "Task"
        ]
      },
      "get": {
        "description": "Returns task details by ID for an authenticated user.",
        "operationId": "TaskController_getTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task fetched successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "User not authenticated."
          },
          "404": {
            "description": "Task not found."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Get task by ID",
        "tags": [
          "Task"
        ]
      },
      "delete": {
        "description": "Deletes a task by ID. Admin only.",
        "operationId": "TaskController_deleteTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Task deleted successfully."
          },
          "401": {
            "description": "User not authenticated."
          },
          "403": {
            "description": "Only admin can delete tasks."
          },
          "404": {
            "description": "Task not found."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Delete task (Admin)",
        "tags": [
          "Task"
        ]
      }
    },
    "/tasks/me": {
      "get": {
        "description": "Updates task details by ID. Admin only.",
        "operationId": "TaskController_getMyTasks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Task updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "User not authenticated."
          },
          "403": {
            "description": "Only admin can update tasks."
          },
          "404": {
            "description": "Task not found."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Update a task (Admin)",
        "tags": [
          "Task"
        ]
      }
    },
    "/tasks/titles": {
      "get": {
        "description": "Returns a paginated list of task titles for quick selection/filtering.",
        "operationId": "TaskController_getAllTaskTitles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Task titles fetched successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTaskTitlesResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "User not authenticated."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Get all task titles",
        "tags": [
          "Task"
        ]
      }
    },
...
  },
  "info": {
    "title": "Task Management System API",
    "description": "A fully functional Task Management System backend built with **NestJS**,\n      providing secure authentication, role-based access control, and task and user management endpoints.\n      ",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "accessToken": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateTaskDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateAssignedTaskStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Status of Tasks",
            "example": "IN_PROGRESS",
            "enum": [
              "NOT_STARTED",
              "IN_PROGRESS",
              "PAUSED",
              "COMPLETED",
              "CANCELED"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {}
      },
      "TaskResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "3a95cdae-9f35-4c98-8ff1-13b9b72123de"
          },
          "title": {
            "type": "string",
            "example": "Fix login bug"
          },
          "content": {
            "type": "object",
            "example": "There is a redirect issue after login.",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "NOT_STARTED",
              "IN_PROGRESS",
              "PAUSED",
              "COMPLETED",
              "CANCELED"
            ],
            "example": "NOT_STARTED"
          },
          "assignedToUserId": {
            "type": "object",
            "example": "b71f183e-9278-4c86-8ce8-748cadaa1e41",
            "nullable": true
          },
          "assignedAt": {
            "type": "object",
            "example": "2025-01-25T10:12:45.000Z",
            "nullable": true
          },
          "completedAt": {
            "type": "object",
            "example": "2025-02-01T16:30:00.000Z",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-01-20T09:15:30.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-01-22T11:45:10.000Z"
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "status",
          "assignedToUserId",
          "assignedAt",
          "completedAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "AssignedUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "USER"
            ],
            "example": "USER"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role",
          "createdAt",
          "updatedAt"
        ]
      },
      "TaskWithAssignedUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "object",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "NOT_STARTED",
              "IN_PROGRESS",
              "PAUSED",
              "COMPLETED",
              "CANCELED"
            ],
            "example": "NOT_STARTED"
          },
          "assignedToUserId": {
            "type": "object",
            "nullable": true
          },
          "assignedToUser": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/AssignedUserDto"
              }
            ]
          },
          "assignedAt": {
            "type": "object",
            "nullable": true
          },
          "completedAt": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "status",
          "assignedToUserId",
          "assignedToUser",
          "assignedAt",
          "completedAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedTasksResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskWithAssignedUserDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
      "TaskTitleDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Fix login bug"
          }
        },
        "required": [
          "title"
        ]
      },
      "PaginatedTaskTitlesResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "List of task titles",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskTitleDto"
            }
          },
          "total": {
            "type": "number",
            "example": 42,
            "description": "Total number of tasks"
          },
          "page": {
            "type": "number",
            "example": 1,
            "description": "Current page number"
          },
          "limit": {
            "type": "number",
            "example": 10,
            "description": "Items per page"
          },
          "totalPages": {
            "type": "number",
            "example": 5,
            "description": "Total number of pages"
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
...
    }
  }
}